refactor some stuff into _yaml file
